Remove events from page, HTML typo fixes
The about section was removed to be properly incorporated in the events & opportunities page. If the contents are needed again, they can be retrieved from the git commit history. There were also typos and formatting issues in the about.html file that included missing closing tags, and improper indentation that made understanding the hierarchy of divs more difficult.